Ads + Python/hcaptcha/pyproject.toml
[project]
name = "hcaptcha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "2captcha-python>=1.5.1",
    "loguru>=0.7.3",
    "playwright>=1.51.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tenacity>=9.1.2",
    "web3>=7.10.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
